# PERF: faster constructors from ea scalars