--- a/pandas/core/dtypes/cast.py
+++ b/pandas/core/dtypes/cast.py
@@ -1652,7 +1652,9 @@ def construct_1d_arraylike_from_scalar(
 
     if isinstance(dtype, ExtensionDtype):
         cls = dtype.construct_array_type()
-        subarr = cls._from_sequence([value] * length, dtype=dtype)
+        subarr = cls._from_sequence([], dtype=dtype)
+        taker = np.broadcast_to(np.intp(-1), length)
+        subarr = subarr.take(taker, allow_fill=True, fill_value=value)
 
     else: